Added SecurityRequestHolderFilter which is registered as the first filter.
It and stores the request and response in the SecurityRequestHolder, which had been done by RequestHolderAuthenticationFilter but later in the chain. RequestHolderAuthenticationFilter also was the entry point for authentication, and was renamed to GrailsUsernamePasswordAuthenticationFilter.

SecurityRequestHolderFilter also wraps the request in a HttpServletRequestWrapper wrapper that is aware of the X-Forwarded-Proto header and returns the correct value from isSecure(), getScheme(), and getServerPort() if the header is present.

This is part of the fix for issue #395