Restored the default value for rejectIfNoRule to true.
Changed BootStrap to not insert initial data since each test suite resets the database.
Created TestRequestmapFilterInvocationDefinition as the 'objectDefinitionSource' bean when using requestmaps, which overrides resetConfigs() in RequestmapFilterInvocationDefinition to add mappings for '/error', '/hack/**', and '/testdata/**' to be able to reset the database and logout.
Added a blankPage() action in HackController which renders nothing and is used as the after-logout url to ensure no resources are requested.
Changed the tests to logout by going directly to the Spring Security filter rather than being redirected by LogoutController.