Issue #521: Informational messages are logged with level "WARN"
I left the `println`s in place because they are there so that developers may see the the current state of the `spring-security-core` plugin when the app starts, regardless of whether they've configured the correct log level.

I've modified all but one of the startup messages to utilize log level `info`.