Grails params missing values with PUT and PATCH after add spring security core
Close: https://github.com/grails-plugins/grails-spring-security-core/issues/554

Close: https://github.com/grails-plugins/grails-spring-security-core/pull/557

PUT requests with content-type `application/x-www-form-urlencoded` were not returning parameters when grails-spring-security-core was enabled.

The reason is that the `HttpPutFormContentFilter` was not registered.  Registering it allows for PUT/PATCH requests to work as expected.

This commit registers the filter and updates existing tests that the change broke.