Fixes #587 switch back to THREADLOCAL as more secure default
The reason for this change back to THREADLOCAL is to provide more security by default with the plugin. Secure by default is a common theme with the Spring Security Core plugin and switching back to THREADLOCAL seems to make sense to align with this theme. It was originally changed with this PR: #517

Also, the Spring Security team has stated that MODE_INHERITABLETHREADLOCAL should not be used in a pooled environment (spring-projects/spring-security#6856) which exemplifies MODE_THREADLOCAL is a better default option for the plugin.